+++ src/pages/notifications.tsx
@@ -4,6 +4,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { supabase } from '../lib/supabaseClient';
 import { useSupabaseAuth } from '../hooks/useSupabaseAuth';
+import { formatRelativeTime } from '../utils/time';
 
 type NotificationRow = {
   id: number;
@@ -41,23 +42,6 @@ function formatNotificationAction(n: NotificationRow): string {
 
   return n.event_type;
 }
-
-function formatRelativeTime(isoString: string): string {
-  const created = new Date(isoString).getTime();
-  const now = Date.now();
-  const diffMs = Math.max(0, now - created);
-
-  const diffMinutes = Math.floor(diffMs / (60 * 1000));
-  if (diffMinutes < 1) return 'just now';
-  if (diffMinutes < 60) return `${diffMinutes} minute${diffMinutes === 1 ? '' : 's'} ago`;
-
-  const diffHours = Math.floor(diffMinutes / 60);
-  if (diffHours < 24) return `${diffHours} hour${diffHours === 1 ? '' : 's'} ago`;
-
-  const diffDays = Math.floor(diffHours / 24);
-  return `${diffDays} day${diffDays === 1 ? '' : 's'} ago`;
-}
-
 
 export default function NotificationsPage() {
   const { user } = useSupabaseAuth();
